Extract skills subsection
